Move JobRunStatus enum and STATUS_MAP to constants module (for re-use in both tools/parser) with new SOURCE_FRESHNESS_STEP_NAME
